3D_Viewer: Make it runnable without the need to start ImageJ.
Create a new class UIAdapter, which receives all calls to the Toolbar,
as well as the calls to showStatus() and showProgress().

When starting the 3D_Viewer, it checks whether the ImageJ window is
visible. If yes, the viewer uses an IJAdapter, which extends UIAdapter
and delegates the methods to ImageJ. If the ImageJ window is not
visible, a NoopAdapter is instantiated, which behaves as if the HAND
tool was selected in ImageJ. Calls to showProgress() etc are ignored.