3D_Viewer: Implement arbitrary lookup tables for individual color channels
This allows to adjust color weights and alpha values of the textures in
an arbitrary way, depending on the input pixel value.

For 8 bit images, all four lookup tables (r, g, b, a) are indexed by
the intensity byte value. For RGB images, the index in the red color
table, e.g., is given by the red byte of the input integer value.
Analogously, the indices are given for the green and blue tables. The
index to the alpha table is given by the resulting average of the
r, g and b channels.

This is a first commit, and only volume renderings are working at the
moment. Surface plots have to be adjusted, as well as isosurfaces.