3D Viewer: avoid animation for frames not shown
When the animation is started for point lists that have just been loaded,
the CPU load goes to 100% because the animation cannot be shown on
hidden frames. Only when the user looks at the frames with landmarks
will the CPU load go down. That's everything but desirable.

Just disable the animation by default.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>