3D_Viewer: Allow to 'saturate' volume renderings.
This is based on Stephan Saalfelds idea how to improve volume
renderings of images containing thin fibres.

Ordinary volume renderings are usually quite dark. This is because
bright structures are usually surrounded by less bright pixels, and
only these darker pixels are seen in volume renderings due to the
alpha value of each pixel being the mean of the single RGB components.

The new 'saturated' volume rendering accounts for this by modifying
the displayed colors: While the alpha value remains the mean
of the single RGB values, these RGB values are rescaled so that
one at least one component is fully saturated (=255). This is done
for every pixel individually.