3D Viewer: Implement shortcuts.
These are not just static shortcuts, but user-adjustable: A small set of
default shortcuts is hardcoded in UniverseSettings, but can be changed
by the user via a shortcut dialog, accessible from the View menu. The
dialog scans the current menu bar and checks for which items a shortcut
exists. Setting a shortcut will save it in UniverseSettings and sets an
accelerator key for the corresponding menu item. User-adjusted shortcuts
will be loaded when the viewer is started, also setting the
corresponding accelerator keys.